+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Develop
+
+- Improve VM execution performance
+
 ## 0.7.0
 
 Initial release